+++ crates/burn-jit/src/kernel/template/mod.rs
@@ -1,0 +1,2 @@
+mod base;
+pub use base::*;
