+++ crates/burn-wgpu/Cargo.toml
@@ -14,6 +14,7 @@ version.workspace = true
 default = ["fusion", "burn-jit/default"]
 fusion = ["burn-fusion", "burn-jit/fusion"]
 autotune = ["burn-jit/autotune"]
+extension = ["burn-jit/extension"]
 doc = ["burn-jit/doc"]
 std = ["burn-jit/std"]
