+++ crates/burn-wgpu/src/lib.rs
@@ -13,16 +13,21 @@ mod runtime;
 #[cfg(feature = "fusion")]
 mod fusion;
 
+#[cfg(feature = "extension")]
+pub use burn_jit::{
+    kernel::{
+        into_contiguous, template::build_info, DynamicKernelSource, SourceTemplate,
+        StaticKernelSource,
+    },
+    kernel_wgsl,
+};
+
 pub use device::*;
 pub use element::*;
 pub use graphics::*;
 pub use runtime::*;
 
 pub use burn_jit::compute::{DynamicKernel, WorkGroup};
-pub use burn_jit::kernel::{
-    build_info, into_contiguous, DynamicKernelSource, SourceTemplate, StaticKernelSource,
-};
-pub use burn_jit::kernel_wgsl;
 pub use burn_jit::{tensor::JitTensor, JitBackend};
 
 #[cfg(feature = "fusion")]
